+++ android/filamat-android/src/main/cpp/MaterialBuilder.cpp
@@ -256,6 +256,13 @@ Java_com_google_android_filament_filamat_MaterialBuilder_nMaterialBuilderShadowM
 }
 
 extern "C" JNIEXPORT void JNICALL
+Java_com_google_android_filament_filamat_MaterialBuilder_nMaterialBuilderTransparentShadow(
+        JNIEnv*, jclass, jlong nativeBuilder, jboolean transparentShadow) {
+    auto builder = (MaterialBuilder*) nativeBuilder;
+    builder->transparentShadow(transparentShadow);
+}
+
+extern "C" JNIEXPORT void JNICALL
 Java_com_google_android_filament_filamat_MaterialBuilder_nMaterialBuilderSpecularAntiAliasing(
         JNIEnv*, jclass, jlong nativeBuilder, jboolean specularAntiAliasing) {
     auto builder = (MaterialBuilder*) nativeBuilder;